+++ lib/cp1161.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2002, 2016 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2002 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,7 +14,8 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, see <http://www.gnu.org/licenses/>.
+ * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
 /*
@@ -43,7 +44,7 @@ static const unsigned short cp1161_2uni[96] = {
 };
 
 static int
-cp1161_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
+cp1161_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 {
   unsigned char c = *s;
   if (c < 0x80) {
@@ -65,7 +66,7 @@ static const unsigned char cp1161_page00[16] = {
 };
 
 static int
-cp1161_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
+cp1161_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 {
   unsigned char c = 0;
   if (wc < 0x0080) {
