+++ lib/cp1252.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2016 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,7 +14,8 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, see <http://www.gnu.org/licenses/>.
+ * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
 /*
@@ -31,7 +32,7 @@ static const unsigned short cp1252_2uni[32] = {
 };
 
 static int
-cp1252_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
+cp1252_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 {
   unsigned char c = *s;
   if (c < 0x80 || c >= 0xa0) {
@@ -75,7 +76,7 @@ static const unsigned char cp1252_page20[48] = {
 };
 
 static int
-cp1252_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
+cp1252_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 {
   unsigned char c = 0;
   if (wc < 0x0080) {
