+++ lib/cp858.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2002, 2016 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2002 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,7 +14,8 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, see <http://www.gnu.org/licenses/>.
+ * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
 /*
@@ -22,7 +23,7 @@
  */
 
 static int
-cp858_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
+cp858_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 {
   unsigned char c = *s;
   if (c < 0x80)
@@ -35,7 +36,7 @@ cp858_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 }
 
 static int
-cp858_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
+cp858_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 {
   unsigned char c = 0;
   if (wc < 0x0080) {
