+++ lib/cp936.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2005, 2016 Free Software Foundation, Inc.
+ * Copyright (C) 2005 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,7 +14,8 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, see <http://www.gnu.org/licenses/>.
+ * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
 /*
@@ -47,7 +48,7 @@
  */
 
 static int
-cp936_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
+cp936_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 {
   /* Try GBK first. */
   {
@@ -89,7 +90,7 @@ cp936_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 }
 
 static int
-cp936_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
+cp936_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 {
   /* Try GBK first. */
   {
