+++ lib/gb12345ext.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2012, 2016 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,7 +14,8 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, see <http://www.gnu.org/licenses/>.
+ * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
 /*
@@ -1003,7 +1004,7 @@ static const unsigned short gb12345ext_2uni_page30[6871] = {
 };
 
 static int
-gb12345ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
+gb12345ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 {
   unsigned char c1 = s[0];
   if ((c1 == 0x21) || (c1 == 0x26) || (c1 == 0x28) || (c1 >= 0x30 && c1 <= 0x79)) {
@@ -1758,7 +1759,7 @@ static const Summary16 gb12345ext_uni2indx_pagefe[5] = {
 };
 
 static int
-gb12345ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
+gb12345ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 {
   if (n >= 2) {
     const Summary16 *summary = NULL;
@@ -1777,9 +1778,9 @@ gb12345ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
       unsigned int i = wc & 0x0f;
       if (used & ((unsigned short) 1 << i)) {
         unsigned short c;
-        /* Keep in 'used' only the bits 0..i-1. */
+        /* Keep in `used' only the bits 0..i-1. */
         used &= ((unsigned short) 1 << i) - 1;
-        /* Add 'summary->indx' and the number of bits set in 'used'. */
+        /* Add `summary->indx' and the number of bits set in `used'. */
         used = (used & 0x5555) + ((used & 0xaaaa) >> 1);
         used = (used & 0x3333) + ((used & 0xcccc) >> 2);
         used = (used & 0x0f0f) + ((used & 0xf0f0) >> 4);
