+++ lib/shift_jisx0213.h
@@ -14,7 +14,8 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, see <http://www.gnu.org/licenses/>.
+ * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
 /*
@@ -39,7 +40,7 @@
 #include "flushwc.h"
 
 static int
-shift_jisx0213_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
+shift_jisx0213_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 {
   ucs4_t last_wc = conv->istate;
   if (last_wc) {
@@ -157,7 +158,7 @@ static const struct { unsigned short base; unsigned short composed; } shift_jisx
 };
 
 static int
-shift_jisx0213_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
+shift_jisx0213_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 {
   int count = 0;
   unsigned short lasttwo = conv->ostate;
@@ -293,7 +294,7 @@ shift_jisx0213_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 }
 
 static int
-shift_jisx0213_reset (conv_t conv, unsigned char *r, size_t n)
+shift_jisx0213_reset (conv_t conv, unsigned char *r, int n)
 {
   state_t lasttwo = conv->ostate;
