+++ lib/utf16le.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2008, 2016 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2008 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,7 +14,8 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, see <http://www.gnu.org/licenses/>.
+ * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
+ * Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
 /*
@@ -24,7 +25,7 @@
 /* Specification: RFC 2781 */
 
 static int
-utf16le_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
+utf16le_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 {
   int count = 0;
   if (n >= 2) {
@@ -51,7 +52,7 @@ ilseq:
 }
 
 static int
-utf16le_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
+utf16le_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 {
   if (!(wc >= 0xd800 && wc < 0xe000)) {
     if (wc < 0x10000) {
